pyproject.toml
[project]
name = "adhoc"
version = "0.1.0"
description = "A package for fetching and processing web content"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.3,<3.0.0",
    "httpx>=0.27.0",
    "markdownify>=0.11.6",
    "readabilipy>=0.3.0",
    "protego>=0.2.1",
    "mcp>=0.1.0",  # Assuming this is a local package
    "langchain>=0.1.0",
    "langchain_openai",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0"
]

[tool.poetry]
packages = [{include = "adhoc", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
